Move on ESPHome External Components
Since the removal of the ESPHome Custom Components in the 2025.2.0
release, this integration no longer works.

So, let's move on External Components which provide a much cleaner and
robust integration with ESPHome & HomeAssistant.
We are taking advantage of this opportunity to expose a native light
effect for the color cycle mode.